+++ nautilus_trader/live/node.py
@@ -80,6 +80,8 @@ class TradingNode:
             save_state=config.save_state,
             loop_sig_callback=self._loop_sig_handler,
             log_level=log_level_from_str(config.log_level.upper()),
+            log_level_file=log_level_from_str(config.log_level_file.upper()),
+            log_file_path=config.log_file_path,
             log_rate_limit=config.log_rate_limit,
         )