+++ tests/integration_tests/adapters/betfair/test_betfair_execution.py
@@ -160,7 +160,6 @@ class TestBaseExecutionClient:
         self.msgbus.subscribe("events.order*", self.events.append)
 
         self.logs = []
-        self.logger.register_sink(handler=self.logs.append)
 
     async def _setup_state(self, order_change_message: Optional[OCM] = None):
         """
@@ -272,6 +271,7 @@ class TestBetfairExecutionClient(TestBaseExecutionClient):
         assert isinstance(updated, OrderUpdated)
         assert updated.price == Price.from_str("0.02000")
 
+    @pytest.mark.skip(reason="Log sinks removed")
     @pytest.mark.asyncio
     async def test_modify_order_error_order_doesnt_exist(self):
         # Arrange
@@ -294,6 +294,7 @@ class TestBetfairExecutionClient(TestBaseExecutionClient):
         ]
         assert logs == expected
 
+    @pytest.mark.skip(reason="Log sinks removed")
     @pytest.mark.asyncio
     async def test_modify_order_error_no_venue_id(self):
         # Arrange
@@ -703,6 +704,7 @@ class TestBetfairExecutionClient(TestBaseExecutionClient):
         assert report.quantity == Quantity(10.0, BETFAIR_QUANTITY_PRECISION)
         assert report.filled_qty == Quantity(0.0, BETFAIR_QUANTITY_PRECISION)
 
+    @pytest.mark.skip(reason="Log sinks removed")
     @pytest.mark.asyncio
     async def test_check_cache_against_order_image(self):
         # Arrange